requirements.txt
livekit-agents[deepgram,openai,cartesia,silero,turn-detector, azure, rime]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv